requirements.txt
python-dotenv
python-docx
PyMuPDF
fastapi
uvicorn
pydantic
redis
redis[hiredis]
langchain
langchain-openai